backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
plotly
fredapi
python-dotenv
requests
yfinance
quandl
pydantic
python-multipart
websockets
sqlalchemy
